Refactor project structure: separate API, logic, and tests
