+++ test/test_card.py
@@ -1,0 +1,4 @@
+from api.index import *
+
+if __name__ == "__main__":
+    app.run(debug=True, host="127.0.0.1", port=8888)
